requirements.txt
aioodbc==0.5.0
aiosqlite==0.21.0
alembic==1.17.0
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
bcrypt==5.0.0
black==25.1.0
certifi==2025.4.26
cffi==2.0.0
cfgv==3.4.0
click==8.1.8
colorama==0.4.6
distlib==0.3.9
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
flake8==7.2.0
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
identify==2.6.10
idna==3.10
iniconfig==2.1.0
isort==6.0.1
Jinja2==3.1.6
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mypy==1.15.0
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.5.0
pre_commit==4.2.0
pyasn1==0.6.1
pycodestyle==2.13.0
pycparser==2.23
pydantic==2.11.4
pydantic_core==2.33.2
pyflakes==3.3.2
Pygments==2.19.1
pyodbc==5.2.0
pytest==8.3.5
pytest-asyncio==0.26.0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.5
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlmodel==0.0.24
starlette==0.46.2
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
virtualenv==20.31.1
watchfiles==1.0.5
websockets==15.0.1
